% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode_physchem}
\alias{encode_physchem}
\title{Encode Aligned Peptide Sequences Using the VHSE8 Matrix}
\usage{
encode_physchem(AA_seqs)
}
\arguments{
\item{AA_seqs}{List of aligned character sequences.}
}
\value{
Rank 3 array of peptide sequence encodings.
}
\description{
Encode Aligned Peptide Sequences Using the VHSE8 Matrix
}
\examples{

examples <- rhoData$data

encode_physchem(examples)

}
