% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{plot_cost_over_rep}
\alias{plot_cost_over_rep}
\title{Plot the Change in Cost Over Gradient Descent Iterations}
\usage{
plot_cost_over_rep(model)
}
\arguments{
\item{model}{A model produced by linear train.}
}
\description{
Visualizes the change in cost over both the training and validation sets
throughout the course of the training. Allows for tuning of hyperparameters
and can help prevent overfitting.
}
\examples{
examples <- rhoData$data
labels <- rhoData$labels

shuffled_datasets <- shuffled_partitions(examples, labels, 650,
                                         encode=encode_physchem)

# Train a linear model to perform regression.


model <- linear_train(shuffled_datasets$e1,
                      shuffled_datasets$l1,
                      shuffled_datasets$e2,
                      shuffled_datasets$l2)


plot_cost_over_rep(model)

}
