% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_train.R
\name{predict}
\alias{predict}
\title{Predict a Continuously Distributed Protein Function}
\usage{
predict(X, model)
}
\arguments{
\item{X}{A rank 3 collection of rank 2 encodings of protein sequences,
encoded under the same scheme as model.}

\item{model}{The model to perform the prediction.}
}
\value{
Vecotr of double predictions.
}
\description{
Performs regression using the given, trained model in order to predict
some continuously ditributed protein function.
}
