% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{property_effect_heatmap}
\alias{property_effect_heatmap}
\title{Heatmap View of Physiochemical Weighings}
\usage{
property_effect_heatmap(model, start = 1, stop = -1)
}
\arguments{
\item{model}{The model as learned by linear_train. This model must have
been trained on a physiochemical encoding.}

\item{start}{The first in the range of residues to be displayed, 1 by
default.}

\item{stop}{The last in the range of residues to be displayed, the last
residue in the alignment by default.}
}
\description{
A function that generates an N x 8 heatmap displaying the weight assigned to
each of the VHSE8 properties across a specified region of the alignment.
Allows for the interpretation of the learned attributes of the dataset. e.g. a red tile at
a given position indicates that physiochemical property is related to an
increase the protein function when present at that position.
}
\details{
Extend produced figure as necceesary for best results.
}
\examples{

examples <- rhoData$data
labels <- rhoData$labels

shuffled_datasets <- shuffled_partitions(examples, labels, 650,
                                         encode=encode_physiochemical)

# Train a linear model to perform regression.

\dontrun{
model <- linear_train(shuffled_datasets$e1,
                      shuffled_datasets$l1,
                      shuffled_datasets$e2,
                      shuffled_datasets$l2)
}

property_effect_heatmap(model)

}
