% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visualise.R
\name{residue_effect_heatmap}
\alias{residue_effect_heatmap}
\title{Heatmap View of Learned One Hot Weights}
\usage{
residue_effect_heatmap(model, start = 1, stop = -1)
}
\arguments{
\item{model}{The model as learned by linear_train. This model must have
been trained on one-hot encoding.}

\item{start}{The first in the range of residues to be displayed, 1 by
default.}

\item{stop}{The last in the range of residues to be displayed, the last
residue in the alignment by default.}
}
\description{
A function that generates an N x 20 heatmap displaying the weight assigned to
each possible residue across a specified region of the alignment. Allows for
the  interpretation of the learned attributes of the dataset. e.g. a red tile
at a given position indicates that amino acid is expected to increase the
protein function when present at that position.
}
\details{
Extend produced figure as necceesary for best results.
}
\examples{

example_data <- get_example_data()

examples <- example_data$data
labels <- example_data$labels

shuffled_datasets <- shuffled_partitions(examples, labels, 650,
                                         encode=encode_onehot)

# Train a linear model to perform regression.

\dontrun{
model <- linear_train(shuffled_datasets$e1,
                      shuffled_datasets$l1,
                      shuffled_datasets$e2,
                      shuffled_datasets$l2)
}

residue_effect_heatmap(model)

}
