% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rhoData}
\alias{rhoData}
\title{Channel Rhodopsins and Their Peak Absorbation Values.}
\format{
An object of class \code{list} of length 2.
}
\source{
(Karasuyama et al., 2018)
}
\usage{
rhoData
}
\description{
The peak absorbation frequencies for a number of channel rhodopsins were
provided by (Karasuyama et al., 2018). We can treat these data as a set of
examples and thier respectives labels. This dataset was produced by aligning
the amino acid sequences of these proteins (using CLUSTAL OMEGA).
}
\details{
This dataset can be substituted for any list of aligned protein sequences and
float values representing some functional property of those proteins.
}
\examples{
\dontrun{
rhoData
}
}
\keyword{datasets}
