% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{shuffled_partitions}
\alias{shuffled_partitions}
\title{Shuffle, Partition, and Encode Datasets.}
\usage{
shuffled_partitions(examples, labels, num_p1, encode = NULL)
}
\arguments{
\item{examples}{A list of aligned amino acid sequences.}

\item{labels}{A list of floating point numbers representing the
experimentally derived protein function for each example sequence.}

\item{num_p1}{The number of examples to put the in the first dataset. The
remainder will be put in the second dataset. This should be greater than 0
and less than or equal to the number of eaxamples in the given dataset.}

\item{encode}{Function to be used to encode the given examples. NULL by
default, returning the original unmodified peptide sequences. The two
encoding functions provided by this package are \code{encode_onehot} and
\code{encode_physchem}.}
}
\description{
Randomly shuffles the given dataset before seperating it into two sets which
are then encoded and ready for use to train a linear model.
}
\examples{
examples <- rhoData$data
labels <- rhoData$labels

# Shuffle, partition, and encode data set.
shuffled_datasets <- shuffled_partitions(examples, labels, 650,
                                         encode=encode_onehot)

train_data <- shuffled_datasets$e1
train_labels <- shuffled_datasets$l1

valid_data <- shuffled_datasets$e2
valid_labels <- shuffled_datasets$l2


}
