% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VHSE}
\alias{VHSE}
\title{The VHSE8 Scale}
\format{
An object of class \code{list} of length 8.
}
\source{
(Mei et al,. 2005)
}
\usage{
VHSE
}
\description{
A list of vectors that allow for the encoding the physiochemical properties
of each amino acid.Commonly used for machine learning in function prediction
tasks.
}
\examples{
\dontrun{
VHSE
}
}
\keyword{datasets}
