% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode_onehot}
\alias{encode_onehot}
\title{Encode Aligned Peptide Sequences into a One-hot Matrix}
\usage{
encode_onehot(AA_seqs)
}
\arguments{
\item{AA_seqs}{List of aligned character sequences.}
}
\value{
Rank 3 array of peptide sequence encodings.
}
\description{
Encode Aligned Peptide Sequences into a One-hot Matrix
}
\examples{

examples <- rhoData$data

encoded_enamples <- encode_onehot(examples)

}
