% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encode.R
\name{encode_physchem}
\alias{encode_physchem}
\title{Encode Aligned Peptide Sequences Using the VHSE8 Matrix}
\usage{
encode_physchem(AA_seqs)
}
\arguments{
\item{AA_seqs}{A list of aligned peptides.Represented as a list of character
vectors.}
}
\value{
Rank 3 array containing all peptide sequence encodings (rank 2).
i.e. an array of peptide sequence encodings.
}
\description{
Encode Aligned Peptide Sequences Using the VHSE8 Matrix
}
\examples{
examples <- rhoData$data

encoded_examples <- encode_physchem(examples)
encoded_examples[1]

}
\references{
Wickham H (2019). assertthat: Easy Pre and Post Assertions.
R package version 0.2.1, https://CRAN.R-project.org/package=assertthat.
}
