% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build.R
\name{maximal_sequence}
\alias{maximal_sequence}
\title{Produce a Sequence to Optimize a Given Function}
\usage{
maximal_sequence(model, do_min = FALSE)
}
\arguments{
\item{model}{The model as learned by linear_train. This model must have
been trained on one-hot encoding.}

\item{do_min}{Whether to change behaviour to instead minimise the protein
function.}
}
\value{
A character representation of the best peptide.
}
\description{
Uses the weights leaned by a model to produce an amino acid sequence that
produces the maximal predited protein function. Will only work on alignments
with low-indel content.
}
\details{
Note that a random base is chose when there is a tie between weights.
}
