% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_train.R
\name{predict}
\alias{predict}
\title{Predict a Continuously Distributed Protein Function}
\usage{
predict(X, model)
}
\arguments{
\item{X}{A rank 3 collection of rank 2 encodings of protein sequences,
encoded under the same scheme as model.}

\item{model}{The model to perform the prediction.}
}
\value{
Vecotr of double predictions.
}
\description{
Performs regression using the given, trained model in order to predict
some continuously ditributed protein function.
}
\examples{

examples <- rhoData$data
labels <- rhoData$labels

shuffled_datasets <- shuffled_partitions(examples, labels, 650,
                                         encode=encode_physiochemical)

# Train a linear model to perform regression.

model <- linear_train(shuffled_datasets$e1,
                      shuffled_datasets$l1,
                      shuffled_datasets$e2,
                      shuffled_datasets$l2)

predict(examples[1, ,], model)

}
