% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_train.R
\name{predictions}
\alias{predictions}
\title{Predict a Continuously Distributed Protein Function}
\usage{
predictions(X, model)
}
\arguments{
\item{X}{Encoded protein sequences. A rank 3 tensor containing examples with
the same dimensionality as those used to train the model.}

\item{model}{The model used to perform the prediction.}
}
\value{
Vector of doubles. Predictions for each input example.
}
\description{
Performs regression using the given, trained model in order to predict
some continuously distributed protein function.
}
\examples{

examples <- rhoData$data
labels <- rhoData$labels

shuffled_datasets <- shuffled_partitions(examples, labels, 650,
                                         encode=encode_physchem)

# Train a linear model to perform regression.

model <- linear_train(shuffled_datasets$e1,
                      shuffled_datasets$l1,
                      shuffled_datasets$e2,
                      shuffled_datasets$l2)

predictions(shuffled_datasets$e1[50:100, ,], model)

}
