% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{rhoData}
\alias{rhoData}
\title{Channel Rhodopsins and Their Peak Absorbation Values.}
\format{
An object of class \code{list} of length 2.
}
\source{
(Bedbrook et al., 2019)
}
\usage{
rhoData
}
\description{
A dataset containing the \lambda max values for several channel rhodopsin
variants as compiled for use in machine learning.
}
\details{
The cooresponding sequences were then aligned using the CLUSTAL OMEGA
alogrithm
}
\examples{
\dontrun{
rhoData
}
}
\keyword{datasets}
