% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VHSE}
\alias{VHSE}
\title{The VHSE Scale}
\format{
An object of class \code{list} of length 8.
}
\source{
(Mei et al,. 2005)
}
\usage{
VHSE
}
\description{
The widely used VHSE scales for encoding protein function.
}
\examples{
\dontrun{
VHSE
}
}
\keyword{datasets}
